Refactor move shared types
