+++ server/packages/server-common/src/utils.ts
@@ -1,4 +1,4 @@
-import { WorkflowInputType } from "./services/requestsDefinitions";
+import { WorkflowInputType } from "../../../../shared/src/requestsDefinitions";
 
 export function isWorkflowInputType(input: string): input is WorkflowInputType {
   return input === "data_input" || input === "data_collection_input";
